# backstage-app/api/openapi.yaml
openapi: 3.0.3
info:
  title: AAP Compliance Backend API
  description: |
    Backend API for the AAP Compliance Pipelines Backstage plugin.
    Provides endpoints for compliance scanning, findings retrieval,
    remediation management, compliance profile registry, and posture reporting.
  version: 0.2.0
  contact:
    name: AAP Compliance Team
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:7007/api/compliance
    description: Local development
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      summary: List available compliance profiles
      description: Returns all compliance profiles available for scanning (DISA STIG, CIS, PCI-DSS, etc.).
      tags:
        - profiles
      parameters:
        - name: framework
          in: query
          description: Filter by compliance framework
          schema:
            type: string
            enum:
              - DISA_STIG
              - CIS
              - PCI_DSS
              - HIPAA
              - NIST_800_53
        - name: os
          in: query
          description: Filter by target operating system
          schema:
            type: string
            example: RHEL 9
      responses:
        '200':
          description: List of compliance profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceProfile'
...
  /workflow-templates:
    get:
      operationId: listWorkflowTemplates
      summary: List compliance workflow templates
      description: Returns workflow job templates from the Controller, optionally filtered by name.
      tags:
        - controller
      parameters:
        - name: name
          in: query
          description: Filter templates by name prefix
          schema:
            type: string
      responses:
        '200':
          description: List of workflow templates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    description:
                      type: string
  /scan:
    post:
      operationId: launchScan
      summary: Launch a compliance scan
      description: Launches a compliance scan against the specified inventory using the selected profile. Dispatches an AAP workflow job template.
      tags:
        - scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequest'
      responses:
        '200':
          description: Scan launched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
        '400':
          description: Invalid request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to launch scan
  /findings:
    get:
      operationId: getFindings
      summary: Get scan findings
      description: Returns findings from a completed compliance scan. If scanId is provided and DB records exist, returns stored findings. Otherwise falls back to the service (mock data in demo mode).
      tags:
        - findings
      parameters:
        - name: scanId
          in: query
          description: Scan ID to retrieve findings for
          schema:
            type: string
      responses:
        '200':
          description: Scan findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MultiHostFinding'
  /workflow-status/{jobId}:
    get:
      operationId: getWorkflowStatus
      summary: Get workflow job status
      description: Returns the current status of a Controller workflow job, used for polling during scan/remediation execution.
      tags:
        - workflow
      parameters:
        - name: jobId
          in: path
          required: true
          description: Controller workflow job ID
          schema:
            type: integer
      responses:
        '200':
          description: Workflow job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowJobStatus'
        '400':
          description: Invalid jobId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /workflow-nodes/{jobId}:
    get:
      operationId: getWorkflowNodes
      summary: Get workflow nodes
      description: Returns the nodes (steps) of a Controller workflow job.
      tags:
        - workflow
      parameters:
        - name: jobId
          in: path
          required: true
          description: Controller workflow job ID
          schema:
            type: integer
      responses:
        '200':
          description: Workflow nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowNode'
        '400':
          description: Invalid jobId
  /job-events/{jobId}:
    get:
      operationId: getJobEvents
      summary: Get job events
      description: Returns events from a Controller job, used for parsing scan output.
      tags:
        - workflow
      parameters:
        - name: jobId
          in: path
          required: true
          description: Controller job ID
          schema:
            type: integer
      responses:
        '200':
          description: Job events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobEvent'
        '400':
          description: Invalid jobId
  /remediate:
    post:
      operationId: applyRemediation
      summary: Apply remediation
      description: |
        Applies remediation for selected rules. Builds an optimized remediation plan
        that groups rules by target host set, then launches one workflow per group.
        Returns the remediation result along with the computed plan.
      tags:
        - remediation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemediationRequest'
      responses:
        '200':
          description: Remediation launched with plan
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RemediationResponse'
                  - type: object
                    properties:
                      plan:
                        $ref: '#/components/schemas/RemediationPlan'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /posture:
    get:
      operationId: getPosture
      summary: Get compliance posture history
      description: Returns historical compliance posture snapshots for trend charts.
      tags:
        - posture
      parameters:
        - name: profileId
          in: query
          description: Filter by compliance profile ID
          schema:
            type: string
        - name: days
          in: query
          description: Number of days of history (default 30)
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Posture history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostureSnapshot'
  /remediation-profiles:
    get:
      operationId: listRemediationProfiles
      summary: List saved remediations
      description: Returns all saved remediations (rule selections, parameters, and scope).
      tags:
        - remediation-profiles
      responses:
        '200':
          description: List of saved remediations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RemediationProfile'
    post:
      operationId: createRemediationProfile
      summary: Save a remediation
      description: Creates a new remediation with the specified rule selections and parameters.
      tags:
        - remediation-profiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRemediationProfileRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemediationProfile'
        '400':
          description: Invalid request
  /cartridges:
    get:
      operationId: listCartridges
      summary: List registered compliance profiles
      description: Returns all registered compliance profiles.
      tags:
        - cartridges
      responses:
        '200':
          description: List of compliance profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceCartridge'
    post:
      operationId: createCartridge
      summary: Register a compliance profile
      description: Creates or updates a compliance profile that maps a standard to automation controller resources.
      tags:
        - cartridges
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCartridgeRequest'
      responses:
        '201':
          description: Compliance profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceCartridge'
        '400':
          description: Invalid request
  /cartridges/{id}:
    delete:
      operationId: deleteCartridge
      summary: Delete a compliance profile
      description: Deletes a registered compliance profile by ID.
      tags:
        - cartridges
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Compliance profile deleted
        '404':
          description: Compliance profile not found
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    ComplianceProfile:
      type: object
      properties:
        id:
          type: string
          example: rhel9-stig
        name:
          type: string
          example: DISA STIG for RHEL 9
        framework:
          type: string
          enum:
            - DISA_STIG
            - CIS
            - PCI_DSS
            - HIPAA
            - NIST_800_53
        version:
          type: string
          example: V2R8
        description:
          type: string
        applicableOs:
          type: array
          items:
            type: string
          example:
            - RHEL 9
        ruleCount:
          type: integer
          example: 366
        lastUpdated:
          type: string
          format: date
        source:
          type: string
          example: ComplianceAsCode/content
    MultiHostFinding:
      type: object
      properties:
        ruleId:
          type: string
          example: sshd_set_idle_timeout
        stigId:
          type: string
          example: V-257844
        title:
          type: string
        description:
          type: string
        fixText:
          type: string
        checkText:
          type: string
        severity:
          type: string
          enum:
            - CAT_I
            - CAT_II
            - CAT_III
        category:
          type: string
        disruption:
          type: string
          enum:
            - low
            - medium
            - high
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              label:
                type: string
              description:
                type: string
              type:
                type: string
                enum: [string, number, boolean, select]
              default:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
              value:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
              options:
                type: array
                items:
                  type: object
                  properties:
                    label:
                      type: string
                    value:
                      oneOf:
                        - type: string
                        - type: number
        hosts:
          type: array
          items:
            $ref: '#/components/schemas/HostFinding'
        passCount:
          type: integer
        failCount:
          type: integer
        totalCount:
          type: integer
    HostFinding:
      type: object
      properties:
        host:
          type: string
        status:
          type: string
          enum:
            - pass
            - fail
            - error
        actualValue:
          type: string
        expectedValue:
          type: string
    WorkflowJobStatus:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        finished:
          type: string
          nullable: true
          format: date-time
        failed:
          type: boolean
        elapsed:
          type: number
        name:
          type: string
    WorkflowNode:
      type: object
      properties:
        id:
          type: integer
        summary_fields:
          type: object
          properties:
            job:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                status:
                  type: string
                type:
                  type: string
        identifier:
          type: string
    JobEvent:
      type: object
      properties:
        id:
          type: integer
        event:
          type: string
        event_data:
          type: object
          additionalProperties: true
        stdout:
          type: string
        host_name:
          type: string
    ScanRequest:
      type: object
      required:
        - profileId
        - inventoryId
      properties:
        profileId:
          type: string
          description: ID of the compliance profile to scan against
          example: rhel9-stig
        inventoryId:
          type: integer
          description: AAP inventory ID
          example: 1
        evaluateOnly:
          type: boolean
          description: If true, scan only without remediation
          default: true
        limit:
          type: string
          description: Host limit pattern (optional)
        workflowTemplateId:
          type: integer
          description: Override workflow template ID (optional)
    ScanResponse:
      type: object
      properties:
        scanId:
          type: string
        workflowJobId:
          type: integer
        status:
          type: string
    RemediationRequest:
      type: object
      required:
        - profileId
        - inventoryId
        - selections
      properties:
        profileId:
          type: string
          description: ID of the compliance profile
        inventoryId:
          type: integer
          description: AAP inventory ID
        selections:
          type: array
          items:
            $ref: '#/components/schemas/RemediationSelection'
        limit:
          type: string
          description: Host limit pattern (optional)
    RemediationResponse:
      type: object
      properties:
        remediationId:
          type: string
        workflowJobId:
          type: integer
        status:
          type: string
    RemediationSelection:
      type: object
      required:
        - ruleId
        - enabled
      properties:
        ruleId:
          type: string
        enabled:
          type: boolean
        parameters:
          type: object
          additionalProperties: true
    RemediationPlan:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/RemediationPlanGroup'
        totalRules:
          type: integer
        totalHosts:
          type: integer
    RemediationPlanGroup:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        limit:
          type: string
        extraVars:
          type: object
          additionalProperties: true
        hostCount:
          type: integer
        ruleCount:
          type: integer
...
    PostureSnapshot:
      type: object
      properties:
        id:
          type: string
        profileId:
          type: string
        timestamp:
          type: string
          format: date-time
        totalHosts:
          type: integer
        totalRules:
          type: integer
        passCount:
          type: integer
        failCount:
          type: integer
        compliancePct:
          type: number
    RemediationProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        complianceProfileId:
          type: string
        targetInventory:
          type: string
        selections:
          type: array
          items:
            $ref: '#/components/schemas/RemediationSelection'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SaveRemediationProfileRequest:
      type: object
      required:
        - name
        - selections
      properties:
        name:
          type: string
        description:
          type: string
        complianceProfileId:
          type: string
        selections:
          type: array
          items:
            $ref: '#/components/schemas/RemediationSelection'
    ComplianceCartridge:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        description:
          type: string
        framework:
          type: string
        version:
          type: string
        platform:
          type: string
        workflowTemplateId:
          type: integer
          nullable: true
        eeId:
          type: integer
          nullable: true
        remediationPlaybookPath:
          type: string
        scanTags:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SaveCartridgeRequest:
      type: object
      required:
        - displayName
        - framework
      properties:
        id:
          type: string
        displayName:
          type: string
        description:
          type: string
        framework:
          type: string
        version:
          type: string
        platform:
          type: string
        workflowTemplateId:
          type: integer
          nullable: true
        eeId:
          type: integer
          nullable: true
        remediationPlaybookPath:
          type: string
        scanTags:
          type: string